requirements.txt
tqdm
jieba
wandb
numpy
pytorch >= 1.6
